refactor(skill-eval): extract Path A/B logic into utils/skill-changes.ts
main.ts is now orchestration only. collectFromYamlChanges and
collectFromMdChanges encapsulate the per-path change detection logic,
and the two paths now run in parallel via Promise.all.
